+++ frontend/src/lib/canvas/ModelCanvas.svelte
@@ -9,7 +9,6 @@
 	import { simpleViewNodeTypes } from './nodes';
 	import { simpleViewEdgeTypes } from './edges';
 	import CanvasAnnouncer from './controls/CanvasAnnouncer.svelte';
-	import CanvasToolbar from './controls/CanvasToolbar.svelte';
 	import KeyboardHandler from './controls/KeyboardHandler.svelte';
 	import type { CanvasNode, CanvasEdge } from '$lib/types/canvas';
 
@@ -141,7 +140,6 @@
 	>
 		<Controls showLock={false} />
 		<Background />
-		<CanvasToolbar onannounce={handleAnnounce} />
 		<KeyboardHandler
 			bind:this={keyboardHandler}
 			{nodes}